Use common processors setup for ABC meta-classes
This fix bug with not supported decorators in controllers
